requirements.txt
# Backend API - Runtime essentials
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.0
cachetools==5.5.0

# Databricks - Runtime only
databricks-sql-connector==3.5.0
databricks-sdk>=0.20.0

# Force reinstall v2 - pip module issue fix
